Function sui_node::metrics::start_metrics_push_task

pub fn start_metrics_push_task(config: &NodeConfig, registry: RegistryService)

Starts a task to periodically push metrics to a configured endpoint if a metrics push endpoint is configured.